+++ CHANGELOG.rst
@@ -10,7 +10,8 @@ Development
 * File for API key is no more useful and commands have been modified to remove the
   ``filekey`` argument;
 * Fixed RST formatter that failed when asked for failures but there was none;
-
+* Added new argument ``--chunk`` on analyze and report commands. It defines the
+  amount of requirements to process before pausing;
 
 
 Version 0.3.0 - 2024/08/27